requirements.txt
asgiref==3.5.2
dj-rest-auth[with_social]==2.2.5
Django==4.1.1
django-allauth==0.50.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.2.0
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
gunicorn==20.1.0
isort==5.12.0
Pillow==9.2.0
PyJWT==2.6.0
pytz==2022.5
python-decouple==3.6
sqlparse==0.4.3
sorl-thumbnail
django-countries
twilio==7.14.0
whitenoise==6.2.0
